Move interfaces to its packages and generate mocks
