+++ mocks/google_cloud_client.go
