+++ mocks/writer_wrapper.go
