+++ pkg/status/interfaces.go
@@ -1,4 +1,4 @@
-package interfaces
+package status
 
 import (
 	"context"
@@ -6,7 +6,6 @@ import (
 	"k8s.io/apimachinery/pkg/runtime"
 	ctrlclient "sigs.k8s.io/controller-runtime/pkg/client"
 )
-
 
 type WriterWrapper interface {
 	Update(ctx context.Context, obj runtime.Object, opts ...ctrlclient.UpdateOption) error
